requirements.txt
cfgv==3.4.0
distlib==0.3.9
dnspython==2.7.0
et_xmlfile==2.0.0
filelock==3.16.1
identify==2.6.1
nodeenv==1.9.1
numpy==2.1.2
openpyxl==3.1.5
pandas==2.2.3
platformdirs==4.3.6
pre_commit==4.0.1
pymongo==4.13.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.2
PyYAML==6.0.2
six==1.16.0
tzdata==2024.2
virtualenv==20.27.0
